+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 # Change Log
 
 ## [Unreleased]
+
+## [0.1.0]
+### Added
+- Run and debug jasmine-based tests
+- CodeLenses for *.spec.ts and *.spec.js files
+- Support for compiled tests
+- Monorepo support
+- Settings in vscode, `jasmine-test-runner.json` and `package.json`